forking : code slightly refactored / some usefull things work now
