+++ Makefile
@@ -7,3 +7,6 @@ deb:
 		--section x11
 	cd deb_dist/obkey-*/; \
 	dpkg-buildpackage -rfakeroot -uc -us
+
+lint:
+	pylint --output-format=parseable --reports=y obkey_classes | tee pylint.log
