+++ misc/obkey.desktop
@@ -1,5 +1,5 @@
 [Desktop Entry]
-Version=1.0
+Version=1.2
 Type=Application
 Name=Openbox Key bindings
 Name[fr]=Raccourcis clavier Openbox
