+++ obkey_parts/__init__.py
@@ -1,0 +1,36 @@
+#!/usr/bin/python2
+"""
+  This file is a part of Openbox Key Editor
+  Copyright (C) 2009 nsf <no.smile.face@gmail.com>
+  v1.1 - Code migrated from PyGTK to PyGObject
+         github.com/stevenhoneyman/obkey
+  v1.2pre1 - solve a minor bug on copy-paste bug
+  v1.2pre2 - 19.06.2016 - structured presentation of actions...
+         github.com/luffah/obkey
+
+  MIT License
+
+  Permission is hereby granted, free of charge, to any person obtaining a copy
+  of this software and associated documentation files (the "Software"), to deal
+  in the Software without restriction, including without limitation the rights
+  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+  copies of the Software, and to permit persons to whom the Software is
+  furnished to do so, subject to the following conditions:
+
+  The above copyright notice and this permission notice shall be included in
+  all copies or substantial portions of the Software.
+
+  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+  THE SOFTWARE.
+"""
+
+from obkey_parts.OpenboxConfig import OpenboxConfig
+from obkey_parts.OBActions import ActionList
+from obkey_parts.KeyTable import KeyTable
+from obkey_parts.PropertyTable import PropertyTable
+from obkey_parts.Gui import Gtk
