+++ po/Makefile
@@ -1,8 +1,8 @@
 LANGS := $(patsubst obkey.%.po,%,$(wildcard *.po))
-TARGETS := $(patsubst %,../locale/%/LC_MESSAGES/obkey.mo,$(LANGS))
+TARGETS := $(patsubst %,../resources/locale/%/LC_MESSAGES/obkey.mo,$(LANGS))
 
 all: $(TARGETS)
 
-../locale/%/LC_MESSAGES/obkey.mo: obkey.%.po
+../resources/locale/%/LC_MESSAGES/obkey.mo: obkey.%.po
 	mkdir -p $(dir $@)
 	msgfmt -o $@ $<
