+++ resources/locale/bs/LC_MESSAGES/obkey.mo
