+++ resources/locale/fr/LC_MESSAGES/obkey.mo
