+++ resources/locale/uk/LC_MESSAGES/obkey.mo
